Improve performance of GPS module
This will improve the performance of the GPS module somewhat. This is
done by giving the GPS full priority when looking for its first fix.
After its first fix its priority is removed, such that the rest of the
app gains precedence. This works because subsequent fixes usually take a
insignificant amount of time:

Signed-off-by: Markus Rekdal <markus.rekdal@nordicsemi.no>