samples: nrf9160: lte_gateway: Remove nrf cloud api references
Previously the LTE Sensor Gateway sample has been hard coded to use nrf
cloud. This will decouple the sample from the nrf cloud api and port it
to the generic cloud api. This means that the sample can be used with
other cloud solutions such as Azure and AWS as well.

Signed-off-by: Markus Rekdal <markus.rekdal@nordicsemi.no>