nrf_security: Key loader for new cc3xx AES
Add an external key loader for AES keys for the
new driver. This is needed because the ARM
cryptocell driver supports AES hardware keys
which we don't support. For example the RTL and
KCP keys.

Since we don't support these keys we need to enable
the option CC3XX_CONFIG_AES_EXTERNAL_KEY_LOADER and
provide an appropriate set_key function.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>